feature(workflow, product) Fixes after CR
