+++ module/product/src/Infrastructure/Strategy/SimpleProductFactoryStrategy.php
@@ -10,19 +10,28 @@ declare(strict_types=1);
 namespace Ergonode\Product\Infrastructure\Strategy;
 
 use Ergonode\Product\Domain\Entity\AbstractProduct;
+use Ergonode\Product\Domain\Entity\SimpleProduct;
 use Ergonode\Product\Domain\ValueObject\Sku;
 use Ergonode\SharedKernel\Domain\Aggregate\ProductId;
 use Ergonode\SharedKernel\Domain\Aggregate\TemplateId;
 
-interface ProductStrategyInterface
+class SimpleProductFactoryStrategy implements ProductFactoryStrategyInterface
 {
-    public function supports(string $type): bool;
+    public function supports(string $type): bool
+    {
+        return SimpleProduct::TYPE === $type;
+    }
 
+    /**
+     * @throws \Exception
+     */
     public function build(
         ProductId $id,
         Sku $sku,
         TemplateId $templateId,
         array $categories,
         array $attributes
-    ): AbstractProduct;
+    ): AbstractProduct {
+        return new SimpleProduct($id, $sku, $templateId, $categories, $attributes);
+    }
 }
