+++ module/workflow/src/Resources/config/services.yml
@@ -50,7 +50,7 @@ services:
         tags:
             - { name: request.param_converter, priority: 2, converter: abstract_workflow_param_converter }
 
-    Ergonode\Workflow\Domain\Factory\ProductFactoryDecorator:
+    Ergonode\Workflow\Domain\Factory\AddStatusToProductFactoryDecorator:
         decorates: Ergonode\Product\Domain\Factory\ProductFactoryInterface
         arguments:
-            - '@Ergonode\Workflow\Domain\Factory\ProductFactoryDecorator.inner'
+            - '@Ergonode\Workflow\Domain\Factory\AddStatusToProductFactoryDecorator.inner'
